Changed edge time from double do int

+++ src/graph/Graph.cpp
@@ -64,7 +64,7 @@ bool Graph::addVertex(int id, const Position &in) {
  * Returns true if successful, and false if the source or destination vertex does not exist.
  */
 
-bool Graph::addEdge(const Position &sourc, const Position &dest, double weight) {
+bool Graph::addEdge(const Position &sourc, const Position &dest, int weight) {
     auto v1 = findVertex(sourc);
     auto v2 = findVertex(dest);
     if (v1 == NULL || v2 == NULL)
@@ -74,7 +74,7 @@ bool Graph::addEdge(const Position &sourc, const Position &dest, double weight) 
     return true;
 }
 
-bool Graph::addEdge(int idNodeOrig, int idNodeDest, double weight) {
+bool Graph::addEdge(int idNodeOrig, int idNodeDest, int weight) {
     Vertex *v1 = this->findVertex(idNodeOrig);
     Vertex *v2 = this->findVertex(idNodeDest);
     if(v1 == NULL || v2 == NULL)
@@ -131,7 +131,7 @@ void Graph::dijkstraShortestPath(Vertex *s) {
     while (!q.empty()) {
         Vertex* v = q.extractMin();
         for (Edge e : v->adj) {
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if (relax(v, e)) {
                 if (oldDist == INF) q.insert(e.dest);
                 else q.decreaseKey(e.dest);
@@ -156,7 +156,7 @@ void Graph::dijkstraShortestPath(Vertex *s, Vertex *d) {
         if (v == d) return;
 
         for (Edge e : v->adj) {
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if (relax(v, e)) {
                 if (oldDist == INF) q.insert(e.dest);
                 else q.decreaseKey(e.dest);
@@ -197,7 +197,7 @@ void Graph::dijkstraShortestPath(Vertex *s, vector<Vertex *> dests) {
         }
 
         for (Edge e : v->adj) {
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if (relax(v, e)) {
                 if (oldDist == INF) q.insert(e.dest);
                 else q.decreaseKey(e.dest);
@@ -230,7 +230,7 @@ Client* Graph::dijkstraClosestClient(Vertex *s, vector<Vertex *> dests) {
         }
 
         for (Edge e : v->adj) {
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if (relax(v, e)) {
                 if (oldDist == INF) q.insert(e.dest);
                 else q.decreaseKey(e.dest);
@@ -247,7 +247,7 @@ Client* Graph::dijkstraClosestClient(Vertex *s, vector<Vertex *> dests) {
  * @param d
  * @return distance between vertices if successful, -1 otherwise
  */
-double Graph::bidirectionalDijkstra(Vertex *s, Vertex *d) {
+int Graph::bidirectionalDijkstra(Vertex *s, Vertex *d) {
     for (Vertex* v : vertexSet) {
         v->dist = INF;
         v->path = nullptr;
@@ -281,7 +281,7 @@ double Graph::bidirectionalDijkstra(Vertex *s, Vertex *d) {
                 return totalDistance;
             }
 
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if(relax(sV, e)) {
                 destV->visited = true;
                 if(oldDist == INF) s_queue.insert(e.dest);
@@ -299,11 +299,11 @@ double Graph::bidirectionalDijkstra(Vertex *s, Vertex *d) {
         for(Edge& e : orderedEdges) {
             Vertex* destV = e.dest;
             if( destV->visited == true){   // If it has already been visited in the other direction
-                double totalDistance = this->joinBidirectionalDistances(destV, dV, e.weight);
+                int totalDistance = this->joinBidirectionalDistances(destV, dV, e.weight);
                 return totalDistance;
             }
 
-            double oldDist = e.dest->dist;
+            int oldDist = e.dest->dist;
             if(relax(dV, e)) {
                 destV->backwardsVisited = true;
                 if(oldDist == INF) d_queue.insert(e.dest);
@@ -314,7 +314,7 @@ double Graph::bidirectionalDijkstra(Vertex *s, Vertex *d) {
     return -1;
 }
 
-double Graph::joinBidirectionalDistances(Vertex *intersectionVertex, Vertex *oppDirectionVertex, double oppDirectionWeight) {
+int Graph::joinBidirectionalDistances(Vertex *intersectionVertex, Vertex *oppDirectionVertex, int oppDirectionWeight) {
     return intersectionVertex->dist + oppDirectionVertex->dist + oppDirectionWeight;
 }
